refactor : change security context to auth context
