+++ src/main/java/coz/weavon/context/auth/application/adapter/AuthMemberAdapter.java
@@ -1,0 +1,9 @@
+package coz.weavon.context.auth.application.adapter;
+
+import coz.weavon.context.auth.domain.model.AuthUser;
+import coz.weavon.context.auth.domain.model.OAuthUser;
+
+public interface AuthMemberAdapter {
+
+    AuthUser findAuthUserAndSaveOAuthUserIfAbsent(OAuthUser oAuthUser);
+}
