+++ src/main/java/coz/weavon/context/auth/application/service/AuthTokenService.java
@@ -1,0 +1,8 @@
+package coz.weavon.context.auth.application.service;
+
+import coz.weavon.context.auth.domain.model.AuthToken;
+import coz.weavon.context.auth.domain.model.AuthUser;
+
+public interface AuthTokenService {
+    AuthToken getAuthTokenByAuthUser(AuthUser authUser);
+}
