+++ src/main/java/coz/weavon/context/auth/domain/model/RegistrationProvider.java
@@ -1,4 +1,4 @@
-package coz.weavon.context.security.domain.model;
+package coz.weavon.context.auth.domain.model;
 
 import coz.weavon.common.application.model.exception.BusinessException;
 import lombok.Getter;
