+++ src/main/java/coz/weavon/context/auth/domain/service/AuthTokenResolver.java
@@ -1,6 +1,6 @@
-package coz.weavon.context.security.domain.service;
+package coz.weavon.context.auth.domain.service;
 
-import coz.weavon.context.security.domain.model.AuthToken;
+import coz.weavon.context.auth.domain.model.AuthToken;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
 import java.util.Date;
