--- src/main/java/coz/weavon/context/security/application/service/AuthTokenService.java
@@ -1,8 +1,0 @@
-package coz.weavon.context.security.application.service;
-
-import coz.weavon.context.security.domain.model.AuthToken;
-import coz.weavon.context.security.domain.model.AuthUser;
-
-public interface AuthTokenService {
-    AuthToken getAuthTokenByAuthUser(AuthUser authUser);
-}
